+++ components/webdriver_server/actions.rs
@@ -279,6 +279,7 @@ impl Handler {
 
         let button = (action.button as u16).into();
         let cmd_msg = WebDriverCommandMsg::MouseButtonAction(
+            session.top_level_browsing_context_id,
             MouseButtonAction::Down,
             button,
             pointer_input_state.x as f32,
@@ -325,6 +326,7 @@ impl Handler {
 
         let button = (action.button as u16).into();
         let cmd_msg = WebDriverCommandMsg::MouseButtonAction(
+            session.top_level_browsing_context_id,
             MouseButtonAction::Up,
             button,
             pointer_input_state.x as f32,
@@ -469,7 +471,11 @@ impl Handler {
             // Step 7
             if x != current_x || y != current_y {
                 // Step 7.2
-                let cmd_msg = WebDriverCommandMsg::MouseMoveAction(x as f32, y as f32);
+                let cmd_msg = WebDriverCommandMsg::MouseMoveAction(
+                    session.top_level_browsing_context_id,
+                    x as f32,
+                    y as f32,
+                );
                 self.constellation_chan
                     .send(ConstellationMsg::WebDriverCommand(cmd_msg))
                     .unwrap();
